+++ src/main/resources/mapper/user/UserMapper.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="com.miracle.coordifit.auth.repository.UserRepository">
+<mapper namespace="com.miracle.coordifit.user.repository.UserRepository">
 
-    <insert id="insertUser" parameterType="com.miracle.coordifit.auth.model.User">
+    <insert id="insertUser" parameterType="com.miracle.coordifit.user.model.User">
         INSERT INTO users (
             user_id, email, password, nickname, file_id,
             login_type_code, kakao_id, gender_code, birth_date, is_active,
@@ -15,7 +15,7 @@
         )
     </insert>
 
-    <select id="selectUserByEmail" parameterType="string" resultType="com.miracle.coordifit.auth.model.User">
+    <select id="selectUserByEmail" parameterType="string" resultType="com.miracle.coordifit.user.model.User">
         SELECT 
             user_id             AS userId,
             email,
@@ -39,7 +39,7 @@
         AND is_active = 'Y'
     </select>
 
-    <select id="selectUserByUserId" parameterType="string" resultType="com.miracle.coordifit.auth.model.User">
+    <select id="selectUserByUserId" parameterType="string" resultType="com.miracle.coordifit.user.model.User">
         SELECT 
             user_id             AS userId,
             email,
@@ -63,7 +63,7 @@
         AND is_active = 'Y'
     </select>
 
-    <select id="selectUserByNickname" parameterType="string" resultType="com.miracle.coordifit.auth.model.User">
+    <select id="selectUserByNickname" parameterType="string" resultType="com.miracle.coordifit.user.model.User">
         SELECT 
             user_id             AS userId,
             email,
@@ -106,7 +106,7 @@
         FROM users
     </select>
 
-    <update id="updateUser" parameterType="com.miracle.coordifit.auth.model.User">
+    <update id="updateUser" parameterType="com.miracle.coordifit.user.model.User">
         UPDATE users 
         SET email = #{email},
             password = #{password},
